Move the `dev-mode` module under the `@bayou` namespace.
